examples/pytorch/requirements.txt
boto3
lightning>=2.0.6
Pillow
torch>=2.0.1
torchdata>=0.6.1
torchvision
webdataset
s3fs
certifi